refact: rename security lib components
